re-org, stage 2
